requirements.txt
pytz==2021.3
sqlparse==0.4.2
psycopg2-binary==2.9.1;
# psycopg2>=2.7,<3.0
gunicorn==20.1.0
wheel==0.37.1
Django==4.0.3
Django==4.0.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
django-cors-headers==3.11.0
gunicorn==20.1.0
Pillow==9.0.1
dj-database-url==0.4.2